+++ ClientApp/components/Layout/Layout.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { NavMenu } from "./NavMenu";
+import { NavMenu } from "../NavMenu/NavMenu";
 import "./Layout.css";
 
 export interface ILayoutProps {
