+++ ClientApp/shared/Interfaces.tsx
